+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/database/AlbumDao.kt
@@ -1,0 +1,20 @@
+package com.igorwojda.showcase.feature.album.data.database
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.igorwojda.showcase.feature.album.data.database.model.AlbumDataEntity
+
+@Dao
+interface AlbumDao {
+
+    @Query("SELECT * FROM albums")
+    suspend fun getAll(): List<AlbumDataEntity>
+
+    @Query("SELECT * FROM albums where artist = :artistName and name = :albumName and mbId = :mbId")
+    suspend fun getAlbum(artistName: String, albumName: String, mbId: String?): AlbumDataEntity
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertAlbums(albums: List<AlbumDataEntity>)
+}
