+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/database/AlbumDatabase.kt
@@ -1,0 +1,11 @@
+package com.igorwojda.showcase.feature.album.data.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import com.igorwojda.showcase.feature.album.data.database.model.AlbumDataEntity
+
+@Database(entities = [AlbumDataEntity::class], version = 1, exportSchema = false)
+abstract class AlbumDatabase : RoomDatabase() {
+
+    abstract fun albums(): AlbumDao
+}
