--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/network/AlbumListNetworkDataModel.kt
@@ -1,5 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.model.network
-
-internal data class AlbumListNetworkDataModel(
-    val album: List<AlbumDataModel>
-)
