--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/model/network/AlbumWikiNetworkDataModel.kt
@@ -1,13 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.model.network
-
-import com.igorwojda.showcase.feature.album.domain.model.AlbumWikiDomainModel
-
-internal data class AlbumWikiDataModel(
-    val published: String,
-    val summary: String
-)
-
-internal fun AlbumWikiDataModel.toDomainModel() = AlbumWikiDomainModel(
-    published = this.published,
-    summary = this.summary
-)
