+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/model/AlbumListNetworkDataModel.kt
@@ -1,0 +1,5 @@
+package com.igorwojda.showcase.feature.album.data.network.model
+
+internal data class AlbumListNetworkDataModel(
+    val album: List<AlbumDataModel>
+)
