+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/model/AlbumWikiNetworkDataModel.kt
@@ -1,0 +1,13 @@
+package com.igorwojda.showcase.feature.album.data.network.model
+
+import com.igorwojda.showcase.feature.album.domain.model.AlbumWikiDomainModel
+
+internal data class AlbumWikiDataModel(
+    val published: String,
+    val summary: String
+)
+
+internal fun AlbumWikiDataModel.toDomainModel() = AlbumWikiDomainModel(
+    published = this.published,
+    summary = this.summary
+)
