+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/retrofit/response/GetAlbumInfoResponse.kt
@@ -1,0 +1,7 @@
+package com.igorwojda.showcase.feature.album.data.network.retrofit.response
+
+import com.igorwojda.showcase.feature.album.data.network.model.AlbumDataModel
+
+internal data class GetAlbumInfoResponse(
+    val album: AlbumDataModel
+)
