+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/retrofit/response/SearchAlbumResponse.kt
@@ -1,0 +1,7 @@
+package com.igorwojda.showcase.feature.album.data.network.retrofit.response
+
+import com.igorwojda.showcase.feature.album.data.network.model.AlbumSearchResultNetworkDataModel
+
+internal data class SearchAlbumResponse(
+    val results: AlbumSearchResultNetworkDataModel
+)
