+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/retrofit/service/AlbumRetrofitService.kt
@@ -1,0 +1,22 @@
+package com.igorwojda.showcase.feature.album.data.network.retrofit.service
+
+import com.igorwojda.showcase.feature.album.data.network.retrofit.response.GetAlbumInfoResponse
+import com.igorwojda.showcase.feature.album.data.network.retrofit.response.SearchAlbumResponse
+import retrofit2.http.POST
+import retrofit2.http.Query
+
+internal interface AlbumRetrofitService {
+
+    @POST("./?method=album.search")
+    suspend fun searchAlbumAsync(
+        @Query("album") phrase: String,
+        @Query("limit") limit: Int = 60
+    ): SearchAlbumResponse
+
+    @POST("./?method=album.getInfo")
+    suspend fun getAlbumInfoAsync(
+        @Query("artist") artistName: String,
+        @Query("album") albumName: String,
+        @Query("mbid") mbId: String?
+    ): GetAlbumInfoResponse?
+}
