--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/room/AlbumDao.kt
@@ -1,20 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.room
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.igorwojda.showcase.feature.album.data.model.database.AlbumDataEntity
-
-@Dao
-interface AlbumDao {
-
-    @Query("SELECT * FROM albums")
-    suspend fun getAll(): List<AlbumDataEntity>
-
-    @Query("SELECT * FROM albums where artist = :artistName and name = :albumName and mbId = :mbId")
-    suspend fun getAlbum(artistName: String, albumName: String, mbId: String?): AlbumDataEntity
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertAlbums(albums: List<AlbumDataEntity>)
-}
