--- feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/room/AlbumDatabase.kt
@@ -1,11 +1,0 @@
-package com.igorwojda.showcase.feature.album.data.room
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import com.igorwojda.showcase.feature.album.data.model.database.AlbumDataEntity
-
-@Database(entities = [AlbumDataEntity::class], version = 1, exportSchema = false)
-abstract class AlbumDatabase : RoomDatabase() {
-
-    abstract fun albums(): AlbumDao
-}
