+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/data/database/model/AlbumDataModelTest.kt
@@ -1,0 +1,73 @@
+package com.igorwojda.showcase.feature.album.data.database.model
+
+import com.igorwojda.showcase.feature.album.data.DataFixtures
+import com.igorwojda.showcase.feature.album.data.network.enum.AlbumDataImageSize
+import com.igorwojda.showcase.feature.album.data.network.model.toDomainModel
+import com.igorwojda.showcase.feature.album.domain.enum.AlbumDomainImageSize
+import com.igorwojda.showcase.feature.album.domain.model.AlbumDomainModel
+import org.amshove.kluent.shouldBeEqualTo
+import org.junit.jupiter.api.Test
+
+class AlbumDataModelTest {
+
+    @Test
+    fun `data model with full data maps to AlbumDomainModel`() {
+        // given
+        val cut = DataFixtures.getAlbum()
+
+        // when
+        val domainModel = cut.toDomainModel()
+
+        // then
+        domainModel shouldBeEqualTo AlbumDomainModel(
+            cut.name,
+            cut.artist,
+            cut.images?.map { it.toDomainModel() } ?: listOf(),
+            cut.wiki?.toDomainModel(),
+            cut.mbId
+        )
+    }
+
+    @Test
+    fun `data model with missing data maps to AlbumDomainModel`() {
+        // given
+        val cut = DataFixtures.getMinimalAlbum()
+
+        // when
+        val domainModel = cut.toDomainModel()
+
+        // then
+        domainModel shouldBeEqualTo AlbumDomainModel(
+            name = "name", artist = "artist", images = emptyList(), wiki = null, mbId = null
+        )
+    }
+
+    @Test
+    fun `mapping filters out unknown size`() {
+        // given
+        val albumDataImages = listOf(AlbumDataImageSize.EXTRA_LARGE, AlbumDataImageSize.UNKNOWN)
+            .map { DataFixtures.getAlbumImage(size = it) }
+        val cut = DataFixtures.getAlbum(images = albumDataImages)
+
+        // when
+        val domainModel = cut.toDomainModel()
+
+        // then
+        domainModel.images.single { it.size == AlbumDomainImageSize.EXTRA_LARGE }
+    }
+
+    @Test
+    fun `mapping filters out blank url`() {
+        // given
+        val albumDataImages = listOf("", "url")
+            .map { DataFixtures.getAlbumImage(url = it) }
+
+        val cut = DataFixtures.getAlbum(images = albumDataImages)
+
+        // when
+        val domainModel = cut.toDomainModel()
+
+        // then
+        domainModel.images.single { it.url == "url" }
+    }
+}
