+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/data/database/model/AlbumWikiDataModelTest.kt
@@ -1,0 +1,24 @@
+package com.igorwojda.showcase.feature.album.data.database.model
+
+import com.igorwojda.showcase.feature.album.data.DataFixtures
+import com.igorwojda.showcase.feature.album.data.network.model.toDomainModel
+import com.igorwojda.showcase.feature.album.domain.model.AlbumWikiDomainModel
+import org.amshove.kluent.shouldBeEqualTo
+import org.junit.jupiter.api.Test
+
+class AlbumWikiDataModelTest {
+
+    @Test
+    fun `maps to AlbumWikiDomainModel`() {
+        // given
+        val published = "published"
+        val summary = "summary"
+        val cut = DataFixtures.getAlbumWikiDataModel(published, summary)
+
+        // when
+        val domainModel = cut.toDomainModel()
+
+        // then
+        domainModel shouldBeEqualTo AlbumWikiDomainModel(published, summary)
+    }
+}
