+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/data/network/enum/AlbumDataImageSizeTest.kt
@@ -1,0 +1,19 @@
+package com.igorwojda.showcase.feature.album.data.network.enum
+
+import org.junit.jupiter.api.Test
+
+class AlbumDataImageSizeTest {
+
+    @Test
+    fun `maps to AlbumDomainImageSize`() {
+        // given
+        val dataEnums = AlbumDataImageSize.values().filterNot { it == AlbumDataImageSize.UNKNOWN }
+
+        // when
+        dataEnums.forEach { it.toDomainEnum() }
+
+        // then
+        // no explicit check is required, because test will crash if any of
+        // the consts in the enums can't be mapped to a domain enum
+    }
+}
