+++ feature_album/src/test/kotlin/com/igorwojda/showcase/feature/album/domain/usecase/GetAlbumListUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.igorwojda.showcase.feature.album.domain.usecase
 
-import com.igorwojda.showcase.feature.album.data.repository.AlbumRepositoryImpl
+import com.igorwojda.showcase.feature.album.data.AlbumRepositoryImpl
 import com.igorwojda.showcase.feature.album.domain.DomainFixtures
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
